Added support for automatically downloading cgminer on OS X (as on Windows)

+++ MultiMiner.Xgminer/Installer.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Diagnostics;
 using System.IO;
 using System.Net;
 using System.Text.RegularExpressions;
@@ -7,23 +8,26 @@ namespace MultiMiner.Xgminer
 {
     public static class Installer
     {
-        private const string cgminerDomain = "ck.kolivas.org";
 
         public static void InstallMiner(MinerBackend minerBackend, string destinationFolder)
         {
             if (minerBackend == MinerBackend.Bfgminer)
                 throw new NotImplementedException();
 
-            string minerUrl = GetLatestCgminerDownloadUrl();
+            //support Windows and OS X for now, we'll go for Linux in the future
+            if (OSVersionPlatform.GetConcretePlatform() == PlatformID.Unix)
+                throw new NotImplementedException();
+
+            string minerUrl = GetCgminerDownloadUrl();
             if (!String.IsNullOrEmpty(minerUrl))
             {
-                string minerDownloadFile = Path.Combine(Environment.GetEnvironmentVariable("TEMP"), "miner.zip");
+                string minerDownloadFile = Path.Combine(Path.GetTempPath(), "miner.zip");
                 File.Delete(minerDownloadFile);
 
                 new WebClient().DownloadFile(new Uri(minerUrl), minerDownloadFile);
                 try
                 {
-                    UnzipFileToFolder(minerDownloadFile, destinationFolder, false, true);
+                    UnzipFileToFolder(minerDownloadFile, destinationFolder);
                 }
                 finally
                 {
@@ -32,9 +36,36 @@ namespace MultiMiner.Xgminer
             }
         }
 
-        private static string GetLatestCgminerDownloadUrl()
+        private static string GetCgminerDownloadUrl()
         {
-            string availableDownloadsHtml = new WebClient().DownloadString("http://" + cgminerDomain + "/apps/cgminer/");
+            if (OSVersionPlatform.GetConcretePlatform() == PlatformID.MacOSX)
+                return GetCgminerMacOSXDownloadUrl();
+            else
+                return GetCgminerWindowsDownloadUrl();
+        }
+
+        public static string GetMinerDownloadRoot(MinerBackend minerBackend)
+        {
+            if (minerBackend == MinerBackend.Bfgminer)
+                throw new NotImplementedException();
+
+            if (OSVersionPlatform.GetConcretePlatform() == PlatformID.MacOSX)
+                return "http://github.com/nwoolls/xgminer-osx";
+            else
+                return "http://ck.kolivas.org";
+        }
+
+        private static string GetCgminerMacOSXDownloadUrl()
+        {
+            //hard-coded for now, dynamic in the future
+            string downloadRoot = GetMinerDownloadRoot(MinerBackend.Cgminer);
+            return String.Format("{0}/releases/download/v1.0.0/cgminer-3.3.1-osx64.tar.gz", downloadRoot);
+        }
+
+        private static string GetCgminerWindowsDownloadUrl()
+        {
+            string downloadRoot = GetMinerDownloadRoot(MinerBackend.Cgminer);
+            string availableDownloadsHtml = new WebClient().DownloadString(String.Format("{0}/apps/cgminer/", downloadRoot));
             const string pattern = @".*<a href=""(cgminer-.+?-windows.zip)";
             Match match = Regex.Match(availableDownloadsHtml, pattern);
             if (match.Success)
@@ -46,9 +77,33 @@ namespace MultiMiner.Xgminer
 
             return "";
         }
-
-        private static void UnzipFileToFolder(string zipFilePath, string destionationFolder, bool showProgress, bool yesToAll)
+        private static void UnzipFileToFolder(string zipFilePath, string destionationFolder)
         {
+            Directory.CreateDirectory(destionationFolder);
+
+            if (OSVersionPlatform.GetGenericPlatform() == PlatformID.Unix)
+                UnzipFileToFolderUnix(zipFilePath, destionationFolder);
+            else
+                UnzipFileToFolderWindows(zipFilePath, destionationFolder);
+        }
+
+        private static void UnzipFileToFolderUnix(string zipFilePath, string destionationFolder)
+        {
+            ProcessStartInfo startInfo = new ProcessStartInfo();
+            startInfo.CreateNoWindow = true;
+            startInfo.WindowStyle = ProcessWindowStyle.Hidden;
+            startInfo.FileName = "tar";
+            startInfo.Arguments = String.Format("-xzvf \"{0}\" -C \"{1}\" --strip-components=1", zipFilePath, destionationFolder);
+            
+            Process process = Process.Start(startInfo);
+            process.WaitForExit();
+        }
+
+        private static void UnzipFileToFolderWindows(string zipFilePath, string destionationFolder)
+        {
+            const bool showProgress = false;
+            const bool yesToAll = true;
+
             Shell32.Folder sourceFolder = GetShell32NameSpace(zipFilePath);
             Directory.CreateDirectory(destionationFolder);
             Shell32.Folder destinationFolder = GetShell32NameSpace(destionationFolder);
